improve: Removed Redundant Testing Stage & Switched TO E2E Tests on Preview Deployments
- Due to the nature of Vercel Preview, it creates preview deployments for all
branches associated to the project. As such, the E2E tests run on all branches for which
Vercel releases a deployment for (including preview) which applies to every branch.

- We will get stability coverage across all branches on all pushes across the project.
As such, it is not necessary to build and execute a web server on a container to E2E
test again. As the build and execute process is automatically tested and reported by
Vercel.